Close #ZZZ - change map for easier pathfinding
Description: The unbreakable obstacles were taking more
than 1 tile sometimes, but in the EnemyTank.java we were
passing just the xy coordinate of the bottom-left tile.
So instead of passing several tiles to the algorithm
we changed the unbreakable objects to be always 1 tile.
Then when we pass the bottom-left xy coordinate to the
pathfinding, it would correctly calculate the paths.